+++ server_utils.h
@@ -1,1 +1,1 @@
-void emit_buffer_to_file(char *file, unsigned char *data, int n_bytes);
+void emit_buffer_to_file(char *file, unsigned char *data, size_t n_bytes);
